+++ examples/src/geometry_viewer.rs
@@ -4,7 +4,7 @@
  * Created Date: 24/05/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 03/06/2023
+ * Last Modified: 27/07/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -18,15 +18,15 @@ fn main() {
     let autd = Controller::builder()
         .add_device(AUTD3::new(Vector3::zeros(), Vector3::zeros()))
         .add_device(AUTD3::new(
-            Vector3::new(0., 0., DEVICE_WIDTH),
+            Vector3::new(0., 0., AUTD3::DEVICE_WIDTH),
             Vector3::new(0., PI / 2., 0.),
         ))
         .add_device(AUTD3::new(
-            Vector3::new(DEVICE_WIDTH, 0., DEVICE_WIDTH),
+            Vector3::new(AUTD3::DEVICE_WIDTH, 0., AUTD3::DEVICE_WIDTH),
             Vector3::new(0., PI, 0.),
         ))
         .add_device(AUTD3::new(
-            Vector3::new(DEVICE_WIDTH, 0., 0.),
+            Vector3::new(AUTD3::DEVICE_WIDTH, 0., 0.),
             Vector3::new(0., -PI / 2., 0.),
         ))
         .open_with(NullLink {})