+++ python/build.ps1
@@ -3,7 +3,7 @@
 # Created Date: 28/05/2023
 # Author: Shun Suzuki
 # -----
-# Last Modified: 28/05/2023
+# Last Modified: 24/07/2023
 # Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 # -----
 # Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -16,10 +16,10 @@ cd ..
 cd capi
 cargo build --release --all
 cd ..
-mkdir python/pyautd3/bin/win_x64 > NUL 2>&1
+mkdir python/pyautd3/bin > NUL 2>&1
 foreach($dll in Get-ChildItem -Path capi/target/release | Where {$_.extension -like ".dll"}){
-    Copy-Item -Path $dll -Destination python/pyautd3/bin/win_x64
+    Copy-Item -Path $dll -Destination python/pyautd3/bin
 }
 cd python
-python setup.py sdist bdist_wheel True
+python -m build -w -C="--build-option=--plat-name" -C="--build-option=win-amd64"
 popd