+++ src/autd3-link-soem/src/local/mod.rs
@@ -4,14 +4,13 @@
  * Created Date: 21/05/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 13/07/2023
+ * Last Modified: 26/07/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
  *
  */
 
-mod ecat;
 mod error;
 mod error_handler;
 mod ethernet_adapters;