+++ src/autd3-modulation-audio-file/src/error.rs
@@ -4,14 +4,13 @@
  * Created Date: 15/06/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 15/06/2023
+ * Last Modified: 23/07/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
  *
  */
 
-use autd3::error::AUTDError;
 use autd3_core::error::AUTDInternalError;
 use thiserror::Error;
 
@@ -35,8 +34,8 @@ impl From<hound::Error> for AudioFileError {
     }
 }
 
-impl From<AudioFileError> for AUTDError {
+impl From<AudioFileError> for AUTDInternalError {
     fn from(value: AudioFileError) -> Self {
-        AUTDError::Internal(AUTDInternalError::ModulationError(value.to_string()))
+        AUTDInternalError::ModulationError(value.to_string())
     }
 }